requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pyjwt==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==0.28.0
google-cloud-vision==3.4.3
PyPDF2==3.0.1
openpyxl==3.1.2